+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## Unreleased
+
+- Fixed an error in document symbol provider when viewing a file located outside of current workspace.
+- Fixed an error in workspace symbol provider when no workspace is open.
+
 ## 2025.9.0
 
 - Added "Rebuild Ctags" editor Command. The old method of building ctags via a Task is going to be deprecated.